requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
backcall=0.2.0=pypi_0
blas=1.0=mkl
bottleneck=1.3.5=py37h7deecbd_0
ca-certificates=2023.01.10=h06a4308_0
certifi=2022.12.7=py37h06a4308_0
charset-normalizer=3.1.0=pypi_0
cycler=0.11.0=pypi_0
debugpy=1.7.0=pypi_0
decorator=5.1.1=pypi_0
entrypoints=0.4=pypi_0
fair-esm=2.0.0=pypi_0
filelock=3.12.2=pypi_0
fonttools=4.38.0=pypi_0
fsspec=2023.1.0=pypi_0
hiddenlayer=0.3=pypi_0
huggingface-hub=0.16.4=pypi_0
idna=3.4=pypi_0
importlib-metadata=6.7.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.16.2=pypi_0
ipython=7.34.0=pypi_0
jedi=0.18.2=pypi_0
joblib=1.2.0=pypi_0
jupyter-client=7.4.9=pypi_0
jupyter-core=4.12.0=pypi_0
kiwisolver=1.4.4=pypi_0
libedit=3.1.20221030=h5eee18b_0
libffi=3.2.1=hf484d3e_1007
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
matplotlib=3.5.3=pypi_0
matplotlib-inline=0.1.6=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.9=pypi_0
numexpr=2.8.4=py37he184ba9_0
numpy=1.21.6=pypi_0
numpy-base=1.21.5=py37ha15fc14_3
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
opencv-python=4.7.0.72=pypi_0
openssl=1.0.2u=h7b6447c_0
packaging=22.0=py37h06a4308_0
pandas=1.3.5=py37h8c16a72_0
parso=0.8.3=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.4.0=pypi_0
pip=22.3.1=py37h06a4308_0
prompt-toolkit=3.0.38=pypi_0
psutil=5.9.5=pypi_0
ptyprocess=0.7.0=pypi_0
py-cpuinfo=9.0.0=pypi_0
pygments=2.15.1=pypi_0
pyparsing=3.0.9=pypi_0
python=3.7.0=h6e4f718_3
python-dateutil=2.8.2=pyhd3eb1b0_0
pytz=2022.7=py37h06a4308_0
pyyaml=6.0=pypi_0
pyzmq=25.1.2=pypi_0
readline=7.0=h7b6447c_5
regex=2024.4.16=pypi_0
requests=2.28.2=pypi_0
safetensors=0.4.5=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=pypi_0
seaborn=0.12.2=pypi_0
setuptools=65.6.3=py37h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.33.0=h62c20be_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.3=pypi_0
torch=1.13.1=pypi_0
torchaudio=0.13.1=pypi_0
torchvision=0.14.1=pypi_0
tornado=6.2=pypi_0
tqdm=4.65.0=pypi_0
traitlets=5.9.0=pypi_0
transformers=4.30.2=pypi_0
typing-extensions=4.5.0=pypi_0
ultralytics=8.0.145=pypi_0
urllib3=1.26.15=pypi_0
wcwidth=0.2.6=pypi_0
wheel=0.38.4=py37h06a4308_0
xz=5.2.10=h5eee18b_1
zipp=3.15.0=pypi_0
zlib=1.2.13=h5eee18b_0
